# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Game Library API
  version: "2.0"
  description: |
    This REST API manages a video game library. Unauthenticated users can access only the /register 
    and /login endpoints. Authentication is handled using HTTP Basic Auth in a stateless manner. Authenticated 
    users can view the list of games, developers, genres, and platforms, as well as get detailed information for each. 
    Users can also view their profile, change their password, or delete their account. Administrators have full access 
    to all entities (games, developers, genres, platforms) and user management. They can create, update, and delete any 
    resource. All data is transferred via DTOs with validation, and appropriate role-based access control is enforced.
    
    To test protected endpoints, click "Authorize" and enter the following credentials:

    — admin (for both login and password)
    — user (for both login and password)
  contact:
    name: Dmitri Kramar
    email: dmitri.kramar@outlook.com
    url: https://dmitri-kramar.github.io/
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    PasswordDTO:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          type: string
          writeOnly: true
          example: oldPassword
        newPassword:
          type: string
          minLength: 4
          maxLength: 25
          writeOnly: true
          example: newPassword
...
    PlatformDTO:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 25
          example: SomePlatform
...
security:
  - basicAuth: [ ]
paths:
...
  /platforms:
    get:
      summary: Get all platforms (available only for AUTHENTICATED users)
      description: Returns a list of all platforms or an empty list if none exist.
      tags: [ Platforms ]
      responses:
        '200':
          description: List of platforms retrieved successfully
        '401':
          description: Unauthorized – Authentication is required

    post:
      summary: Add a new platform (available only for ADMIN)
      description: Adds a new platform to the database.
      tags: [ Platforms ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlatformDTO'
            examples:
              validPlatform:
                summary: Example of a valid platform
                value:
                  name: Nintendo Switch
      responses:
        '201':
          description: Platform added successfully
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized – Authentication is required
        '403':
          description: Forbidden – Access denied
  /platforms/{id}:
    get:
      summary: Get a platform by ID (available only for AUTHENTICATED users)
      description: Returns a platform by ID if it exists.
      tags: [ Platforms ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the platform
      responses:
        '200':
          description: Platform retrieved successfully
        '401':
          description: Unauthorized – Authentication is required
        '404':
          description: Platform with the given ID was not found

    put:
      summary: Update a platform name (available only for ADMIN)
      description: Updates the platform's name using the given ID.
      tags: [ Platforms ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlatformDTO'
            examples:
              updatePlatform:
                summary: Example update
                value:
                  name: Wii
      responses:
        '200':
          description: Platform updated successfully
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized – Authentication is required
        '403':
          description: Forbidden – Access denied
        '404':
          description: Platform with the given ID was not found

    delete:
      summary: Delete a platform (available only for ADMIN)
      description: Deletes the specified platform by ID.
      tags: [ Platforms ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the platform to delete
      responses:
        '204':
          description: Platform deleted successfully
        '401':
          description: Unauthorized – Authentication is required
        '403':
          description: Forbidden – Access denied
        '404':
          description: Platform with the given ID was not found
...
  /users:
    get:
      summary: Get all users (available only for ADMIN)
      description: Retrieves a list of all users.
      tags: [ Users ]
      responses:
        '200':
          description: Successfully retrieved the list of users
        '401':
          description: Unauthorized – Authentication is required
        '403':
          description: Forbidden – ADMIN role is required
  /users/{id}:
    get:
      summary: Get a user by ID (available only for AUTHENTICATED users)
      description: Returns a user profile by ID.
      tags: [ Users ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the user to retrieve
      responses:
        '200':
          description: User retrieved successfully
        '401':
          description: Unauthorized – Authentication is required
        '404':
          description: User with the given ID was not found

    put:
      summary: Update user password (available only for the user himself)
      description: Updates the user's password. Requires the correct old password.
      tags: [ Users ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the user whose password will be changed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordDTO'
            examples:
              updatePassword:
                summary: Example password update
                value:
                  oldPassword: "user"
                  newPassword: "user123"
      responses:
        '200':
          description: Password updated successfully
        '400':
          description: Invalid input or incorrect old password
        '401':
          description: Unauthorized – Authentication is required
        '403':
          description: Forbidden – Only the user himself or admin can update

    delete:
      summary: Delete a user (available only for ADMIN or the user himself)
      description: Deletes the user with the given ID.
      tags: [ Users ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the user to delete
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized – Authentication is required
        '403':
          description: Forbidden – Only the user himself or admin can delete
        '404':
          description: User with the given ID was not found